feat - 응답된 사용자 리스트 받아오는 api 연동 완료, 이전 테스트 api 삭제

+++ domain/src/main/java/com/yapp/growth/domain/entity/TimeTableUnit.kt
@@ -4,5 +4,5 @@ data class TimeTableUnit(
     val index: Int,
     val count: Int,
     val users: List<User>,
-    val color: Int
+    val color: Long
 )
